New functionality: Download category_matrix
Deleted "id" from input for functions run_search and run_lecat_analysis as it wasn't in use. But added a line to add IDs to the results files if the corpus contains an ID column (which it should, as the sample file specifies ID as the first column)
Added a function "create_category_matrix" which outputs a file that can be pasted to the corpus and adds columns which permit to use filter and sort to explore the results of the analysis (at the level of category)
Modified documentation as appropriate and improved a few bits and bobs.